% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissCqN-package.R
\docType{package}
\name{dissCqN-package}
\alias{dissCqN-package}
\alias{_PACKAGE}
\title{dissCqN: Multiple Assemblage Dissimilarity for Orders q = 0-N}
\description{
Calculate multiple or pairwise dissimilarity for orders q = 0-N 
    (CqN) for a set of species assemblages or interaction networks. See Jost et 
    al. (2011, in ISBN:978-0-19-958067-5).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://murphymv.github.io/dissCqN/}
  \item \url{https://github.com/murphymv/dissCqN}
  \item Report bugs at \url{https://github.com/murphymv/dissCqN/issues}
}

}
\author{
\strong{Maintainer}: Mark Murphy \email{murphymv@gmail.com}

}
\keyword{internal}
