% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissCqN-fun.R
\name{dissCqN}
\alias{dissCqN}
\title{Multiple Assemblage Dissimilarity}
\usage{
dissCqN(mat, q = 0:2, shared.spp = FALSE)
}
\arguments{
\item{mat}{A matrix with assemblages in rows and species or species
interactions in columns. Alternatively, a list of matrices, which will be
interpreted as interaction networks and used to construct an assemblage x
interaction matrix.}

\item{q}{Integer, the orders of \code{q} for which to calculate
dissimilarity. Can be any set of integers between 0 and \emph{N} (the
number of assemblages in \code{mat}).}

\item{shared.spp}{Logical, whether to compare networks of shared species only
(if \code{mat} is a list of networks).}
}
\value{
A numeric vector of dissimilarities for the orders of \code{q}.
}
\description{
Multiple assemblage dissimilarity for orders \emph{q} =
  0-\emph{N}.
}
\references{
Chao, A., Jost, L., Chiang, S. C., Jiang, Y.-H., & Chazdon, R. L.
  (2008). A Two-Stage Probabilistic Approach to Multiple-Community Similarity
  Indices. \emph{Biometrics}, \strong{64}(4), 1178–1186.
  \url{https://doi.org/fcvn63}

  Diserud, O. H., & Ødegaard, F. (2007). A multiple-site similarity measure.
  \emph{Biology Letters}, \strong{3}(1), 20–22. \url{https://doi.org/bwhfx6}

  Jost, L., Chao, A., & Chazdon, R. L. (2011). Compositional similarity and
  beta diversity. In A. E. Magurran & B. J. McGill (Eds.), \emph{Biological
  Diversity: Frontiers in Measurement and Assessment} (pp. 66–84). Oxford
  University Press.
}
