% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissCqN-fun.R
\name{dissMat}
\alias{dissMat}
\title{Pairwise Dissimilarity Matrix}
\usage{
dissMat(
  mat,
  q = 0,
  compare.sub = NULL,
  shared.spp = FALSE,
  parallel = "no",
  ncpus = NULL,
  cl = NULL
)
}
\arguments{
\item{mat}{A matrix with assemblages in rows and species or species
interactions in columns. Alternatively, a list of matrices, which will be
interpreted as interaction networks and used to construct an assemblage x
interaction matrix.}

\item{q}{Integer, the order of \code{q} for which to calculate dissimilarity
(passed to \code{\link[dissCqN]{dissCqN}}). Should be a single integer
value, and if not, only the first value will be used.}

\item{compare.sub}{Subsets of assemblages to compare pairwise. These should
be supplied as a list of two sets of assemblage names or indices. If only
one set is supplied, this set is compared to all other assemblages in
\code{mat}. If more than two sets are supplied, only the first two are
used. If \code{NULL}, all assemblages are compared.}

\item{shared.spp}{Logical, whether to compare networks of shared species only
(if \code{mat} is a list of networks).}

\item{parallel}{The type of parallel processing to use, if any. Can be one of
\code{"snow"}, \code{"multicore"}, or \code{"no"} (for none - the default).}

\item{ncpus}{Number of system cores to use for parallel processing. If
\code{NULL} (default), all available cores are used.}

\item{cl}{Optional cluster to use if \code{parallel = "snow"}. If \code{NULL}
(default), a local cluster is created using the specified number of cores.}
}
\value{
A matrix of pairwise assemblage dissimilarities.
}
\description{
Generate pairwise dissimilarity matrix from a set of species
  assemblages.
}
