% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissCqN-fun.R
\name{intMat}
\alias{intMat}
\title{Assemblage x Species Interaction Matrix}
\usage{
intMat(net, shared.spp = FALSE, ...)
}
\arguments{
\item{net}{An interaction network, or list of networks, supplied as matrices.}

\item{shared.spp}{Logical, whether to use networks of shared species only.}

\item{...}{Arguments to \code{\link[=netShared]{netShared()}} (if \code{shared.spp = TRUE}).}
}
\value{
A matrix with assemblages in rows and species interactions in
columns.
}
\description{
Generate a matrix of assemblages x species interactions from a
set of networks.
}
