% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissCqN-package.R
\docType{package}
\name{dissCqN-package}
\alias{dissCqN-package}
\alias{_PACKAGE}
\title{dissCqN: Multiple Assemblage Dissimilarity for Orders q = 0-N}
\description{
Calculate multiple or pairwise dissimilarity for orders q = 0-N (CqN; Chao et al. 2008 <doi:10/fcvn63>) for a set of species assemblages or interaction networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://murphymv.github.io/dissCqN/}
  \item \url{https://github.com/murphymv/dissCqN}
  \item Report bugs at \url{https://github.com/murphymv/dissCqN/issues}
}

}
\author{
\strong{Maintainer}: Mark Murphy \email{murphymv@gmail.com}

}
\keyword{internal}
